requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
decorator==5.1.1
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docassemble-backports==1.0
executing==2.1.0
GDAL @ D:\v1\v00.1\Authentification-system-with-django-main\GDAL-3.4.3-cp311-cp311-win_amd64.whl
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipdb==0.13.13
ipython==8.27.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
zope.interface==7.0.3
